requirements.txt
python-dotenv>=1.0.0
google-genai 
Pillow>=10.0.0
streamlit>=1.30.0
agno>=0.1.0
duckduckgo-search>=4.1.1
